refactor: 데이터 레이어 개선 - 목 데이터 통합 및 Context API 준비
- SITE_CONFIG로 모든 사이트 설정 통합 (author, bio, skills)
- AboutPage props 간소화 (threads만 받도록)
- JSX 주석 전체 제거 (32개)
- .clauderules 추가 (JSX 주석 금지, 데이터 관리 규칙)
- Context API 마이그레이션 준비
  - SiteConfigProvider 생성
  - useSiteConfig 훅 생성
  - 작업 계획서 추가 (docs/CONTEXT_API_MIGRATION.md)
