+++ src/shared/providers/index.ts
@@ -1,0 +1,1 @@
+export { SiteConfigProvider, useSiteConfig } from './SiteConfigProvider';
